avaluaproy/requirements4.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2
django-allauth==0.60.0
django-crispy-forms==2.1
django-filter==23.5
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.6
importlib-metadata==7.0.1
Markdown==3.5.2
oauthlib==3.2.2
Pillow==10.1.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
typing-extensions==4.8.0
urllib3==2.1.0
zipp==3.17.0
